Simplified members table style derivation
ref https://linear.app/ghost/issue/BER-3470/member-list-fails-to-render-in-chrome-for-massive-member-counts
Moved members table style generation next to the layout math so the list no longer rebuilds repetitive column style objects by hand.
